refactor: move shared prefix helpers from prefix.go into diff.go
Move matchesPrefix and normalizePrefixes to the bottom of diff.go and
their tests to diff_test.go. Drop duplicate tests from include_test.go
and exclude_test.go. Delete prefix.go.
